+++ x/evm/keeper/querier.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"fmt"
 	"strconv"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -51,7 +52,7 @@ func queryProtocolVersion(keeper Keeper) ([]byte, error) {
 
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, hexutil.Uint(vers))
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -64,7 +65,7 @@ func queryBalance(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error)
 	res := types.QueryResBalance{Balance: utils.MarshalBigInt(balance)}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -75,7 +76,7 @@ func queryBlockNumber(ctx sdk.Context, keeper Keeper) ([]byte, error) {
 	bnRes := types.QueryResBlockNumber{Number: num}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, bnRes)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -88,7 +89,7 @@ func queryStorage(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error)
 	res := types.QueryResStorage{Value: val.Bytes()}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 	return bz, nil
 }
@@ -99,7 +100,7 @@ func queryCode(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error) {
 	res := types.QueryResCode{Code: code}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -111,7 +112,7 @@ func queryNonce(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error) {
 	nRes := types.QueryResNonce{Nonce: nonce}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, nRes)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -124,7 +125,7 @@ func queryHashToHeight(ctx sdk.Context, path []string, keeper Keeper) ([]byte, e
 	res := types.QueryResBlockNumber{Number: blockNumber}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -133,18 +134,18 @@ func queryHashToHeight(ctx sdk.Context, path []string, keeper Keeper) ([]byte, e
 func queryBlockLogsBloom(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error) {
 	num, err := strconv.ParseInt(path[1], 10, 64)
 	if err != nil {
-		panic("could not unmarshall block number: " + err.Error())
+		return nil, fmt.Errorf("could not unmarshall block number: %s" + err.Error())
 	}
 
 	bloom, err := keeper.GetBlockBloomMapping(ctx, num)
 	if err != nil {
-		panic("failed to get block bloom mapping: " + err.Error())
+		return nil, fmt.Errorf("failed to get block bloom mapping: %s" + err.Error())
 	}
 
 	res := types.QueryBloomFilter{Bloom: bloom}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -154,14 +155,13 @@ func queryTxLogs(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error) 
 	txHash := ethcmn.HexToHash(path[1])
 	logs, err := keeper.GetLogs(ctx, txHash)
 	if err != nil {
-		// TODO: handle err
-		return nil, nil
+		return nil, err
 	}
 
 	res := types.QueryETHLogs{Logs: logs}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 
 	return bz, nil
@@ -173,7 +173,7 @@ func queryLogs(ctx sdk.Context, keeper Keeper) ([]byte, error) {
 	res := types.QueryETHLogs{Logs: logs}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 	return bz, nil
 }
@@ -189,7 +189,7 @@ func queryAccount(ctx sdk.Context, path []string, keeper Keeper) ([]byte, error)
 	}
 	bz, err := codec.MarshalJSONIndent(keeper.cdc, res)
 	if err != nil {
-		panic(sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error()))
+		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
 	}
 	return bz, nil
 }
