+++ x/evm/types/key.go
@@ -14,3 +14,14 @@ const (
 	// RouterKey uses module name for routing
 	RouterKey = ModuleName
 )
+
+var bloomPrefix = []byte("bloom")
+var logsPrefix = []byte("logs")
+
+func BloomKey(key []byte) []byte {
+	return append(bloomPrefix, key...)
+}
+
+func LogsKey(key []byte) []byte {
+	return append(logsPrefix, key...)
+}
